Restructuring and refactoring
