+++ src/components/ui-kit/ControlButton.js
@@ -1,7 +1,7 @@
+import { memo } from "react";
 import classNames from "classnames";
 import PropTypes from "prop-types";
 import './ControlButton.scss';
-import { memo } from "react";
 
 export const controlTypes = {
   play: 'play',
