+++ src/components/ui-kit/CustomButton.scss
@@ -1,4 +1,4 @@
-@import "../variables.css";
+@import "../../variables.css";
 
 .Button {
   color: var(--controls);
