+++ src/components/ui-kit/Volume.scss
@@ -1,4 +1,4 @@
-@import "../variables.css";
+@import "../../variables.css";
 
 .VolumeBar{
   &[type="range"] {
