--- src/legacy/css/player.css
@@ -1,250 +1,0 @@
-/* Default Skin */
-body {
-	background: #eefaff;
-}
-audio {
-	width: 100%;
-}
-h3 {
-	margin: 0 0 0.5em 0;
-	color: #777;
-	white-space: nowrap;
-}
-.top-row>div {
-	padding-top: 1em;
-}
-.no_padd{
-	padding-left: 0;
-	padding-right: 0;
-}
-.orange{
-	color: #F97534;
-}
-.light .orangeBG{
-	background-color: #F97534;
-	color: black !important;
-}
-.dark .orangeBG{
-	background-color: #F97534;
-	color: rgb(192,192,192) !important;
-}
-
-/*player*/
-.player{
-	padding: 15px;
-	overflow: auto;
-	width: 350px;
-	margin: 0 auto;
-	margin-bottom: 20px;
-}
-
-.player>div{
-	overflow: auto;
-	min-height: 50px;
-	float: left;
-	margin:0 10px;
-}
-#NameStation{
-	display: block;
-	width: 100%;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-}
-
-/*Кнопки вперед и назад*/
-div.mw_control{
-	width: 40%;
-}
-.bottom-np{
-	width: auto;
-	height: 25px;
-	float: left;
-}
-.contrl-button{
-	height: 100%;
-	width:25px;
-	float:left;	
-	background-size:100% 100%;
-	cursor: pointer;
-	-webkit-transition:all .3s;
-	-moz-transition:all .3s;
-	-ms-transition:all .3s;
-	transition:all .3s;
-}
-.bottom-np .next{
-	background-image: url("../img/next.png");
-}
-.bottom-np .prev{
-	background-image: url("../img/prev.png");
-	margin-right: 10px;
-}
-
-/* play/stop */
-div.mw_play{
-	width: 5%;
-	min-width: 50px;
-}
-#playStop{
-	width: 50px;
-	height: 50px;
-	border:1px solid #F97534;
-	border-radius:0;
-	background-size: 100% 100%;
-	background-repeat: no-repeat;
-	-webkit-transition:all .5s;
-	-moz-transition:all .5s;
-	-ms-transition:all .5s;
-	transition:all .5s;
-}
-#playStop.play{
-	background-size: 100% 100%;
-	background-image: url("../img/play.png");
-}
-#playStop.pause{
-	background-size: 100% 100%;
-	background-image: url("../img/pause.png");
-	border-radius:50%;
-}
-#playStop:hover{
-	border-radius:50%;
-	cursor:pointer;
-}
-#playStop:active, .contrl-button:active, #volume:active, .mw_switcher:active, #btn_streams_list:active{
-	-webkit-transform: scale(.9,.9);
-	-moz-transform: scale(.9,.9);
-	-ms-transform: scale(.9,.9);
-	transform: scale(.9,.9);
-}
-
-/*Звук*/
-div.mw_volume{
-	min-width: 80px;
-}
-#volume{
-	height: 25px;
-	width: 25px;
-	float: left;
-	cursor: pointer;
-	background-image: url('../img/volume_b.png');
-	background-size: 100% 100%;
-	background-repeat: no-repeat;
-	-webkit-transition:all .5s;
-	-moz-transition:all .5s;
-	-ms-transition:all .5s;
-	transition:all .5s;
-}
-#volume+span{
-	cursor: default;
-}
-#volume[on=false]{
-	background-image: url('../img/volume_bn.png');
-}
-div.ui-slider-horizontal{
-	margin: 15px 0 0 0;
-}
-/*Фон слайдера*/
-.slider {
-width:80px;
-height:10px;
-position:relative;
-margin:0;
-padding:0 10px;
-border:1px solid #F97534;
-}
-/*Стиль кнопки слайдера*/
-.ui-slider-handle {
-}
-/*Полоска заполнителя слайдера*/
-.ui-widget-header {
-background: url(../img/volume_full.png) no-repeat left;
-background-size: 100% 100%;
-height: 100%;
-left: 0.1px;
-position: absolute;
-}
-a {
-	outline:none;
-	-moz-outline-style:none;
-}
-
-
-/*Список станций*/
-.mw_streams_list{
-	width: 30%;
-}
-#btn_streams_list{
-	border: 1px solid #F97534;
-	margin-top: 10px;
-	height: 30px;
-	width: 100%;
-	text-align: center;
-	line-height: 30px;
-	cursor: pointer;
-	-webkit-transition:all .5s;
-	-moz-transition:all .5s;
-	-ms-transition:all .5s;
-	transition:all .5s;
-	}
-div#drop_down_streams{
-	width: 100%;
-	margin: 0 !important;
-	min-height: 0;
-	height: 220px;
-	overflow: hidden;
-}
-#drop_down_streams>h3{
-	border-top: 1px solid #777;
-}
-#drop_down_streams>ol{
-	overflow: auto;
-	height: 175px;
-}
-#drop_down_streams>ol>li{
-	cursor: pointer;
-}
-#drop_down_streams>ol>li:hover{
-	color: #F97534;
-}
-
-
-/* Tune theme switcher */
-div.mw_switcher{
-	width: 30%;
-	min-height: 32px !important;
-	margin: 10px 10px 0 10px !important;
-	float: right !important;
-	text-align: center;
-	cursor: pointer;
-	-webkit-transition:all .5s;
-	-moz-transition:all .5s;
-	-ms-transition:all .5s;
-	transition:all .5s;
-}
-div[data-role="theme-switcher"]{
-	width: 100%;
-	height: 30px;
-	line-height: 30px;
-	border:1px solid #F97534;
-}
-
-.dark [data-role="theme-switcher"] button {
-	background: #036;
-	color: #779;
-}
-
-.dark [data-role="theme-switcher"] button.active {
-	background: #000;
-}
-/* Dark */
-.player.dark{
-	background: rgba(4,4,4,.8);
-	color: rgb(192,192,192);
-}
-/* Light */
-.player.light{
-	background: rgba(192,192,192,.8);
-}
-
-
-
