--- src/legacy/index.html
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<title>HTML5 Audio Player</title>
-		<meta name="viewport" content="initial-scale=1.0">
-		<meta charset="utf-8">
-		<link rel="stylesheet" type="text/css" href="css/bootstrap.min.css">
-		<script type="text/javascript" src="js/lib/jquery-2.1.0.min.js"></script>
-		<script type="text/javascript" src="js/lib/bootstrap.min.js"></script>
-		<!-- Player -->
-		<script type="text/javascript" src="js/app.js"></script>
-		<link rel="stylesheet" type="text/css" href="css/player.css">
-		<!-- For slider -->		
-		<script src="http://ajax.googleapis.com/ajax/libs/jqueryui/1/jquery-ui.min.js"></script>
-	</head>
-	<body>
-		<div class="container-fluid" data-role="wrapper">
-			<div class="row top-row">
-				<div class="col-md-6 col-sm-6 col-xs-12">
-					<h3 class="pull-left">HTML5 Audio Player</h3>
-				</div>
-				<div class="col-md-6 col-sm-6 col-xs-12 no_padd">
-					<!-- TODO: Audio player here please -->
-					<div class="col-md-12 col-sm-12 col-xs-12 no_padd">
-						<div id="player" class="player light">
-							<div class="mw_play">
-								<div id="playStop" class="play pull-left" title="Play/Pause"></div>
-							</div>
-							<div class="mw_control">
-								<span id="NameStation"></span>
-								<div class="bottom-np">							
-									<div id="prev" class="prev contrl-button" title="Previous station"></div>
-									<div id="next" class="next contrl-button" title="Next station"></div>
-								</div>
-							</div>
-							<div class="mw_volume">
-								<div id="volume" on="true" title="Volume on/off"></div>
-								<span class="orange">Volume</span>
-								<div class="slider"></div>
-							</div>
-							<div class="mw_streams_list">
-								<div title="Streams list" id="btn_streams_list" class="orange" on="true">Streams</div>
-							</div>
-							<div class="mw_switcher">
-								<div title="Color theme" class="orange" data-role="theme-switcher" themes="light">Dark</div>
-							</div>
-							<div id="drop_down_streams">
-								<h3>Streams list</h3>
-								<ol></ol>
-							</div>
-						</div>						
-					</div>	
-				</div>
-			</div>
-		</div>
-	</body>
-</html>
