+++ src/pages/Discovery.scss
@@ -1,4 +1,4 @@
-@import '../mixins';
+@import 'mixins';
 
 .Discovery {
   @include theme-change-styles;
