+++ src/pages/PlayerContainer.scss
@@ -1,4 +1,4 @@
-@import '../variables.css';
+@import 'variables.css';
 .Player__container {
   height: 100%;
   width: 100%;
